+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-08-13 14:22:23
-// Commit: docs(security): update readme and API contracts for security
+// Updated at: 2026-08-16 09:54:05
+// Commit: chore(security): update dependencies and cleanup unused imports in security
 
 class securityComponentState {
-  final String commitRef = '3576';
-  final String lastMessage = 'docs(security): update readme and API contracts for security';
-  final DateTime timestamp = DateTime.parse('2026-08-13T14:22:23+05:00');
+  final String commitRef = '3623';
+  final String lastMessage = 'chore(security): update dependencies and cleanup unused imports in security';
+  final DateTime timestamp = DateTime.parse('2026-08-16T09:54:05+05:00');
 }